requirements.txt
fastapi
uvicorn[standard]
bleak
scapy
pyyaml
websockets
psutil
python-dotenv